feat: refactoring h2h mint
