+++ src/features/gateway/steps/flows/MintH2H.tsx
@@ -18,7 +18,6 @@ import { BigTopWrapper } from "../../../../components/layout/LayoutHelpers";
 import { PaperContent } from "../../../../components/layout/Paper";
 import { Debug } from "../../../../components/utils/Debug";
 import { paths } from "../../../../pages/routes";
-import { useNotifications } from "../../../../providers/Notifications";
 import { getAssetConfig } from "../../../../utils/assetsConfig";
 import { decimalsAmount } from "../../../../utils/numbers";
 import { trimAddress } from "../../../../utils/strings";
@@ -52,6 +51,7 @@ import {
   useChainTransactionSubmitter,
   usePartialTxMemo,
   useRenVMChainTransactionStatusUpdater,
+  useTxRecovery,
 } from "../../gatewayTransactionHooks";
 import { parseGatewayQueryString } from "../../gatewayUtils";
 import {
@@ -162,58 +162,16 @@ export const MintH2HGatewayProcess: FunctionComponent<
   );
   useSetGatewayContext(gateway);
 
-  // TODO: DRY
-  const { showNotification } = useNotifications();
   const [recoveryMode] = useState(hasRenVMHash || hasPartialTx);
-  const [recoveringStarted, setRecoveringStarted] = useState(false);
-  const [recoveringError, setRecoveringError] = useState<Error | null>(null);
-  const { persistLocalTx, findLocalTx } = useTxsStorage();
+  const { persistLocalTx } = useTxsStorage();
 
-  useEffect(() => {
-    // TODO: add partialTx recovery notification
-    if (
-      recoveryMode &&
-      renVMHash &&
-      fromAddress &&
-      gateway !== null &&
-      !recoveringStarted
-    ) {
-      setRecoveringStarted(true);
-      console.log("recovering tx: " + trimAddress(renVMHash));
-      const localTx = findLocalTx(fromAddress, renVMHash);
-      if (localTx === null) {
-        console.error(`Unable to find tx for ${fromAddress}, ${renVMHash}`);
-        return;
-      } else {
-        recoverLocalTx(renVMHash, localTx)
-          .then(() => {
-            showNotification(
-              `Transaction ${trimAddress(renVMHash)} recovered.`,
-              {
-                variant: "success",
-              }
-            );
-          })
-          .catch((error) => {
-            console.error(`Recovering error`, error.message);
-            showNotification(`Failed to recover transaction`, {
-              variant: "error",
-            });
-            setRecoveringError(error);
-          });
-      }
-    }
-  }, [
-    recoveryMode,
-    showNotification,
-    fromAddress,
-    toAddress,
-    renVMHash,
-    recoveringStarted,
-    findLocalTx,
+  const { recoveringError } = useTxRecovery({
     gateway,
+    renVMHash,
+    fromAddress,
+    recoveryMode,
     recoverLocalTx,
-  ]);
+  });
 
   const transaction = transactions[0] || null;
   (window as any).gateway = gateway;
